+++ src/message.ts
@@ -19,6 +19,49 @@ export type Message = {
   isfuzzy: boolean;
 };
 
+function nextMessagWithCondition(
+  document: vscode.TextDocument,
+  lineno: number,
+  condition: Function,
+  backwards = false
+): Message {
+  let message = currentMessageDefinition(document, lineno);
+  const getMessage = backwards ? previousMessage : nextMessage;
+  while (message !== null) {
+    message = getMessage(document, message);
+    if (message && condition(message)) {
+      return message;
+    }
+  }
+  return null;
+}
+
+export function nextUntranslatedMessage(
+  document: vscode.TextDocument,
+  lineno: number,
+  backwards = false
+): Message {
+  return nextMessagWithCondition(
+    document,
+    lineno,
+    (message) => !message.msgstr,
+    backwards
+  );
+}
+
+export function nextFuzzyMessage(
+  document: vscode.TextDocument,
+  lineno: number,
+  backwards = false
+): Message {
+  return nextMessagWithCondition(
+    document,
+    lineno,
+    (message) => message.isfuzzy,
+    backwards
+  );
+}
+
 export function nextMessage(
   document: vscode.TextDocument,
   currentMessage: Message
